booking wip

+++ holidaymaker_java_rest_server/src/main/java/com/example/demo/repository/CustomerRepository.java
@@ -5,8 +5,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
   //  @Query("select c from Customer c where c.email =?1")
     Customer findByEmail(String email);
